ui changes + restore apps when terminate

+++ Flow Work/Flow Work/Views/LobbyView.swift
@@ -96,34 +96,40 @@ struct JoinSessionView: View {
                 .onSubmit {
                     viewModel.joinSession(viewModel.joinSessionCode)
                 }
-            List(viewModel.sessionState.availableSessions) { session in
-                HStack {
-                    SessionItem(session: session)
-                        .foregroundColor(selectedSessionId == session.id ? Color.white : Color.primary)
-                        .padding(.init(top: 5, leading: 10, bottom: 5, trailing: 10))
-                        .contentShape(Rectangle())
-                        .background(selectedSessionId == session.id ? Color.blue : Color.clear)
-                        .gesture(TapGesture(count: 2).onEnded {
-                            viewModel.joinSession(session.id)
-                        })
-                        .simultaneousGesture(TapGesture().onEnded {
-                            selectedSessionId = session.id
-                        })
-                        .contextMenu {
-                            Button(action: {
-                                if (viewModel.authState.isSignedIn) {
-                                    self.viewModel.storeService.removeUserFromSession(userId: viewModel.authState.currentUser!.id, sessionId: session.id)
+            if viewModel.isSessionListLoading {
+                Spacer()
+                ProgressView()
+                Spacer()
+            } else {
+                List(viewModel.sessionState.availableSessions) { session in
+                    HStack {
+                        SessionItem(session: session)
+                            .foregroundColor(selectedSessionId == session.id ? Color.white : Color.primary)
+                            .padding(.init(top: 5, leading: 10, bottom: 5, trailing: 10))
+                            .contentShape(Rectangle())
+                            .background(selectedSessionId == session.id ? Color.blue : Color.clear)
+                            .gesture(TapGesture(count: 2).onEnded {
+                                viewModel.joinSession(session.id)
+                            })
+                            .simultaneousGesture(TapGesture().onEnded {
+                                selectedSessionId = session.id
+                            })
+                            .contextMenu {
+                                Button(action: {
+                                    if (viewModel.authState.isSignedIn) {
+                                        self.viewModel.storeService.removeUserFromSession(userId: viewModel.authState.currentUser!.id, sessionId: session.id)
+                                    }
+                                }) {
+                                    Text("Remove")
                                 }
-                            }) {
-                                Text("Remove")
                             }
-                        }
+                    }
                 }
+                .padding(0)
+                .frame(minHeight: 200)
+                .cornerRadius(5)
+                .listStyle(.plain)
             }
-            .padding(0)
-            .frame(minHeight: 200)
-            .cornerRadius(5)
-            .listStyle(.plain)
             
             HStack(spacing: 10) {
                 Button(action: {
